EZP-25088: move ES field value mappers to Common namespace
