+++ eZ/Publish/Core/Search/Common/FieldValueMapper/FloatMapper.php
@@ -1,23 +1,23 @@
 <?php
 
 /**
- * File containing the BooleanMapper document field value mapper class.
+ * This file is part of the eZ Publish Kernel package.
  *
  * @copyright Copyright (C) eZ Systems AS. All rights reserved.
  * @license For full copyright and license information view LICENSE file distributed with this source code.
  *
  * @version //autogentag//
  */
-namespace eZ\Publish\Core\Search\Elasticsearch\Content\FieldValueMapper;
+namespace eZ\Publish\Core\Search\Common\FieldValueMapper;
 
-use eZ\Publish\Core\Search\Elasticsearch\Content\FieldValueMapper;
-use eZ\Publish\SPI\Search\FieldType\BooleanField;
+use eZ\Publish\Core\Search\Common\FieldValueMapper;
+use eZ\Publish\SPI\Search\FieldType\FloatField;
 use eZ\Publish\SPI\Search\Field;
 
 /**
- * Maps BooleanField document field values to something Elasticsearch can index.
+ * Maps FloatField document field values to something Elasticsearch can index.
  */
-class BooleanMapper extends FieldValueMapper
+class FloatMapper extends FieldValueMapper
 {
     /**
      * Check if field can be mapped.
@@ -28,7 +28,7 @@ class BooleanMapper extends FieldValueMapper
      */
     public function canMap(Field $field)
     {
-        return $field->type instanceof BooleanField;
+        return $field->type instanceof FloatField;
     }
 
     /**
@@ -40,6 +40,6 @@ class BooleanMapper extends FieldValueMapper
      */
     public function map(Field $field)
     {
-        return (boolean)$field->value;
+        return (float)$field->value;
     }
 }
