+++ eZ/Publish/Core/Search/Common/FieldValueMapper/GeoLocationMapper.php
@@ -1,22 +1,23 @@
 <?php
 
 /**
- * File containing the MultipleIdentifierMapper document field value mapper class.
+ * File containing the GeoLocationMapper document field value mapper class.
  *
  * @copyright Copyright (C) eZ Systems AS. All rights reserved.
  * @license For full copyright and license information view LICENSE file distributed with this source code.
  *
  * @version //autogentag//
  */
-namespace eZ\Publish\Core\Search\Elasticsearch\Content\FieldValueMapper;
+namespace eZ\Publish\Core\Search\Common\FieldValueMapper;
 
-use eZ\Publish\SPI\Search\FieldType\MultipleIdentifierField;
+use eZ\Publish\Core\Search\Common\FieldValueMapper;
+use eZ\Publish\SPI\Search\FieldType\GeoLocationField;
 use eZ\Publish\SPI\Search\Field;
 
 /**
- * Maps MultipleIdentifierField document field values to something Elasticsearch can index.
+ * Maps GeoLocationField document field values to something Elasticsearch can index.
  */
-class MultipleIdentifierMapper extends IdentifierMapper
+class GeoLocationMapper extends FieldValueMapper
 {
     /**
      * Check if field can be mapped.
@@ -27,7 +28,7 @@ class MultipleIdentifierMapper extends IdentifierMapper
      */
     public function canMap(Field $field)
     {
-        return $field->type instanceof MultipleIdentifierField;
+        return $field->type instanceof GeoLocationField;
     }
 
     /**
@@ -39,12 +40,9 @@ class MultipleIdentifierMapper extends IdentifierMapper
      */
     public function map(Field $field)
     {
-        $values = array();
-
-        foreach ($field->value as $value) {
-            $values[] = $this->convert($value);
-        }
-
-        return $values;
+        return array(
+            'lat' => $field->value['latitude'],
+            'lon' => $field->value['longitude'],
+        );
     }
 }
