+++ eZ/Publish/Core/Search/Common/FieldValueMapper/MultipleIntegerMapper.php
@@ -1,41 +1,39 @@
 <?php
 
 /**
- * File containing the MultipleStringMapper document field value mapper class.
+ * File containing the MultipleStringMapper class.
  *
  * @copyright Copyright (C) eZ Systems AS. All rights reserved.
  * @license For full copyright and license information view LICENSE file distributed with this source code.
  *
  * @version //autogentag//
  */
-namespace eZ\Publish\Core\Search\Elasticsearch\Content\FieldValueMapper;
+namespace eZ\Publish\Core\Search\Common\FieldValueMapper;
 
-use eZ\Publish\SPI\Search\FieldType;
 use eZ\Publish\SPI\Search\Field;
+use eZ\Publish\SPI\Search\FieldType\MultipleIntegerField;
 
 /**
- * Maps MultipleStringField document field values to something Elasticsearch can index.
+ * Maps raw document field values to something Solr can index.
  */
-class MultipleStringMapper extends StringMapper
+class MultipleIntegerMapper extends IntegerMapper
 {
     /**
      * Check if field can be mapped.
      *
-     * @param \eZ\Publish\SPI\Search\Field $field
+     * @param Field $field
      *
      * @return bool
      */
     public function canMap(Field $field)
     {
-        return
-            $field->type instanceof FieldType\MultipleStringField ||
-            $field->type instanceof FieldType\FullTextField;
+        return $field->type instanceof MultipleIntegerField;
     }
 
     /**
-     * Map field value to a proper Elasticsearch representation.
+     * Map field value to a proper Solr representation.
      *
-     * @param \eZ\Publish\SPI\Search\Field $field
+     * @param Field $field
      *
      * @return array
      */
