+++ eZ/Publish/Core/Search/Common/FieldValueMapper/MultipleStringMapper.php
@@ -1,23 +1,22 @@
 <?php
 
 /**
- * File containing the MultipleBooleanMapper document field value mapper class.
+ * File containing the MultipleStringMapper document field value mapper class.
  *
  * @copyright Copyright (C) eZ Systems AS. All rights reserved.
  * @license For full copyright and license information view LICENSE file distributed with this source code.
  *
  * @version //autogentag//
  */
-namespace eZ\Publish\Core\Search\Elasticsearch\Content\FieldValueMapper;
+namespace eZ\Publish\Core\Search\Common\FieldValueMapper;
 
-use eZ\Publish\Core\Search\Elasticsearch\Content\FieldValueMapper;
-use eZ\Publish\SPI\Search\FieldType\MultipleBooleanField;
+use eZ\Publish\SPI\Search\FieldType;
 use eZ\Publish\SPI\Search\Field;
 
 /**
- * Maps MultipleBooleanField document field values to something Elasticsearch can index.
+ * Maps MultipleStringField document field values to something Elasticsearch can index.
  */
-class MultipleBooleanMapper extends FieldValueMapper
+class MultipleStringMapper extends StringMapper
 {
     /**
      * Check if field can be mapped.
@@ -28,7 +27,9 @@ class MultipleBooleanMapper extends FieldValueMapper
      */
     public function canMap(Field $field)
     {
-        return $field->type instanceof MultipleBooleanField;
+        return
+            $field->type instanceof FieldType\MultipleStringField ||
+            $field->type instanceof FieldType\FullTextField;
     }
 
     /**
@@ -36,14 +37,14 @@ class MultipleBooleanMapper extends FieldValueMapper
      *
      * @param \eZ\Publish\SPI\Search\Field $field
      *
-     * @return mixed
+     * @return array
      */
     public function map(Field $field)
     {
         $values = array();
 
         foreach ((array)$field->value as $value) {
-            $values[] = (boolean)$value;
+            $values[] = $this->convert($value);
         }
 
         return $values;
