+++ eZ/Publish/Core/settings/search_engines/elasticsearch.yml
@@ -3,7 +3,6 @@ imports:
     - {resource: search_engines/elasticsearch/criterion_visitors_content.yml}
     - {resource: search_engines/elasticsearch/criterion_visitors_location.yml}
     - {resource: search_engines/elasticsearch/facet_builder_visitors.yml}
-    - {resource: search_engines/elasticsearch/field_value_mappers.yml}
     - {resource: search_engines/elasticsearch/services.yml}
     - {resource: search_engines/elasticsearch/sort_clause_visitors_content.yml}
     - {resource: search_engines/elasticsearch/sort_clause_visitors_location.yml}
@@ -24,7 +23,7 @@ services:
     ezpublish.search.elasticsearch.serializer:
         class: %ezpublish.search.elasticsearch.serializer.class%
         arguments:
-            - @ezpublish.search.elasticsearch.content.field_value_mapper.aggregate
+            - @ezpublish.search.common.field_value_mapper.aggregate
             - @ezpublish.search.common.field_name_generator
 
     ezpublish.search.elasticsearch.mapper.standard:
