--- eZ/Publish/Core/settings/search_engines/elasticsearch/field_value_mappers.yml
@@ -1,80 +1,0 @@
-parameters:
-    ezpublish.search.elasticsearch.content.field_value_mapper.boolean.class: eZ\Publish\Core\Search\Elasticsearch\Content\FieldValueMapper\BooleanMapper
-    ezpublish.search.elasticsearch.content.field_value_mapper.date.class: eZ\Publish\Core\Search\Elasticsearch\Content\FieldValueMapper\DateMapper
-    ezpublish.search.elasticsearch.content.field_value_mapper.document.class: eZ\Publish\Core\Search\Elasticsearch\Content\FieldValueMapper\DocumentMapper
-    ezpublish.search.elasticsearch.content.field_value_mapper.geo_location.class: eZ\Publish\Core\Search\Elasticsearch\Content\FieldValueMapper\GeoLocationMapper
-    ezpublish.search.elasticsearch.content.field_value_mapper.float.class: eZ\Publish\Core\Search\Elasticsearch\Content\FieldValueMapper\FloatMapper
-    ezpublish.search.elasticsearch.content.field_value_mapper.identifier.class: eZ\Publish\Core\Search\Elasticsearch\Content\FieldValueMapper\IdentifierMapper
-    ezpublish.search.elasticsearch.content.field_value_mapper.integer.class: eZ\Publish\Core\Search\Elasticsearch\Content\FieldValueMapper\IntegerMapper
-    ezpublish.search.elasticsearch.content.field_value_mapper.multiple_integer.class: eZ\Publish\Core\Search\Elasticsearch\Content\FieldValueMapper\MultipleIntegerMapper
-    ezpublish.search.elasticsearch.content.field_value_mapper.multiple_boolean.class: eZ\Publish\Core\Search\Elasticsearch\Content\FieldValueMapper\MultipleBooleanMapper
-    ezpublish.search.elasticsearch.content.field_value_mapper.multiple_identifier.class: eZ\Publish\Core\Search\Elasticsearch\Content\FieldValueMapper\MultipleIdentifierMapper
-    ezpublish.search.elasticsearch.content.field_value_mapper.multiple_string.class: eZ\Publish\Core\Search\Elasticsearch\Content\FieldValueMapper\MultipleStringMapper
-    ezpublish.search.elasticsearch.content.field_value_mapper.price.class: eZ\Publish\Core\Search\Elasticsearch\Content\FieldValueMapper\PriceMapper
-    ezpublish.search.elasticsearch.content.field_value_mapper.string.class: eZ\Publish\Core\Search\Elasticsearch\Content\FieldValueMapper\StringMapper
-
-services:
-    ezpublish.search.elasticsearch.content.field_value_mapper.boolean:
-        class: %ezpublish.search.elasticsearch.content.field_value_mapper.boolean.class%
-        tags:
-            - {name: ezpublish.search.elasticsearch.content.field_value_mapper}
-
-    ezpublish.search.elasticsearch.content.field_value_mapper.date:
-        class: %ezpublish.search.elasticsearch.content.field_value_mapper.date.class%
-        tags:
-            - {name: ezpublish.search.elasticsearch.content.field_value_mapper}
-
-    ezpublish.search.elasticsearch.content.field_value_mapper.document:
-        class: %ezpublish.search.elasticsearch.content.field_value_mapper.document.class%
-        tags:
-            - {name: ezpublish.search.elasticsearch.content.field_value_mapper}
-
-    ezpublish.search.elasticsearch.content.field_value_mapper.geo_location:
-        class: %ezpublish.search.elasticsearch.content.field_value_mapper.geo_location.class%
-        tags:
-            - {name: ezpublish.search.elasticsearch.content.field_value_mapper}
-
-    ezpublish.search.elasticsearch.content.field_value_mapper.float:
-        class: %ezpublish.search.elasticsearch.content.field_value_mapper.float.class%
-        tags:
-            - {name: ezpublish.search.elasticsearch.content.field_value_mapper}
-
-    ezpublish.search.elasticsearch.content.field_value_mapper.identifier:
-        class: %ezpublish.search.elasticsearch.content.field_value_mapper.identifier.class%
-        tags:
-            - {name: ezpublish.search.elasticsearch.content.field_value_mapper}
-
-    ezpublish.search.elasticsearch.content.field_value_mapper.multiple_integer:
-        class: %ezpublish.search.elasticsearch.content.field_value_mapper.multiple_integer.class%
-        tags:
-            - {name: ezpublish.search.elasticsearch.content.field_value_mapper}
-
-    ezpublish.search.elasticsearch.content.field_value_mapper.integer:
-        class: %ezpublish.search.elasticsearch.content.field_value_mapper.integer.class%
-        tags:
-            - {name: ezpublish.search.elasticsearch.content.field_value_mapper}
-
-    ezpublish.search.elasticsearch.content.field_value_mapper.multiple_boolean:
-        class: %ezpublish.search.elasticsearch.content.field_value_mapper.multiple_boolean.class%
-        tags:
-            - {name: ezpublish.search.elasticsearch.content.field_value_mapper}
-
-    ezpublish.search.elasticsearch.content.field_value_mapper.multiple_identifier:
-        class: %ezpublish.search.elasticsearch.content.field_value_mapper.multiple_identifier.class%
-        tags:
-            - {name: ezpublish.search.elasticsearch.content.field_value_mapper}
-
-    ezpublish.search.elasticsearch.content.field_value_mapper.multiple_string:
-        class: %ezpublish.search.elasticsearch.content.field_value_mapper.multiple_string.class%
-        tags:
-            - {name: ezpublish.search.elasticsearch.content.field_value_mapper}
-
-    ezpublish.search.elasticsearch.content.field_value_mapper.price:
-        class: %ezpublish.search.elasticsearch.content.field_value_mapper.price.class%
-        tags:
-            - {name: ezpublish.search.elasticsearch.content.field_value_mapper}
-
-    ezpublish.search.elasticsearch.content.field_value_mapper.string:
-        class: %ezpublish.search.elasticsearch.content.field_value_mapper.string.class%
-        tags:
-            - {name: ezpublish.search.elasticsearch.content.field_value_mapper}
